+++ packages/next/src/shared/lib/router/utils/route-regex.ts
@@ -5,6 +5,7 @@ import {
 import { INTERCEPTION_ROUTE_MARKERS } from './interception-routes'
 import { escapeStringRegexp } from '../../escape-regexp'
 import { removeTrailingSlash } from './remove-trailing-slash'
+import { PARAMETER_PATTERN, parseMatchedParameter } from './get-dynamic-param'
 
 export interface Group {
   pos: number
@@ -75,62 +76,6 @@ type GetRouteRegexOptions = {
    * Whether to exclude the optional trailing slash from the route regex.
    */
   excludeOptionalTrailingSlash?: boolean
-}
-
-/**
- * Regular expression pattern used to match route parameters.
- * Matches both single parameters and parameter groups.
- * Examples:
- *   - `[[...slug]]` matches parameter group with key 'slug', repeat: true, optional: true
- *   - `[...slug]` matches parameter group with key 'slug', repeat: true, optional: false
- *   - `[[foo]]` matches parameter with key 'foo', repeat: false, optional: true
- *   - `[bar]` matches parameter with key 'bar', repeat: false, optional: false
- */
-const PARAMETER_PATTERN = /^([^[]*)\[((?:\[[^\]]*\])|[^\]]+)\](.*)$/
-
-/**
- * Parses a given parameter from a route to a data structure that can be used
- * to generate the parametrized route.
- * Examples:
- *   - `[[...slug]]` -> `{ key: 'slug', repeat: true, optional: true }`
- *   - `[...slug]` -> `{ key: 'slug', repeat: true, optional: false }`
- *   - `[[foo]]` -> `{ key: 'foo', repeat: false, optional: true }`
- *   - `[bar]` -> `{ key: 'bar', repeat: false, optional: false }`
- *   - `fizz` -> `{ key: 'fizz', repeat: false, optional: false }`
- * @param param - The parameter to parse.
- * @returns The parsed parameter as a data structure.
- */
-export function parseParameter(param: string) {
-  const match = param.match(PARAMETER_PATTERN)
-
-  if (!match) {
-    return parseMatchedParameter(param)
-  }
-
-  return parseMatchedParameter(match[2])
-}
-
-/**
- * Parses a matched parameter from the PARAMETER_PATTERN regex to a data structure that can be used
- * to generate the parametrized route.
- * Examples:
- *   - `[...slug]` -> `{ key: 'slug', repeat: true, optional: true }`
- *   - `...slug` -> `{ key: 'slug', repeat: true, optional: false }`
- *   - `[foo]` -> `{ key: 'foo', repeat: false, optional: true }`
- *   - `bar` -> `{ key: 'bar', repeat: false, optional: false }`
- * @param param - The matched parameter to parse.
- * @returns The parsed parameter as a data structure.
- */
-function parseMatchedParameter(param: string) {
-  const optional = param.startsWith('[') && param.endsWith(']')
-  if (optional) {
-    param = param.slice(1, -1)
-  }
-  const repeat = param.startsWith('...')
-  if (repeat) {
-    param = param.slice(3)
-  }
-  return { key: param, repeat, optional }
 }
 
 function getParametrizedRoute(
